Improvements to behaviour in IE
I have changed "submit" to be an ordinary button.

I traced through all the errors in IE many were down to trying to assign
to value of an output element.  IE does not allow this you have to
assign to innerHtml (which also works in Chrome) instead.

There is still one fault I haven't traced in IE which is calling .value
on an object that doesn't have one (in IE).  The error appears to be
benign.