Licence page improvements
the create licence popup now works correctly, showing the user the
created licence, and then giving them the option to copy it to the
clipboard. Came across a problem in which the html5 'required' global
attribute no longer works, as the submit button on the form has been
changed from a submit to a button (i assume this is the problem anyway)
so I have created code which now works around this problem, the user can
no longer create a licence with no username and an invalid date.